tests/submission/PhilipTaylor/canvas/size.attributes.removed.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: size.attributes.removed</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.size.html">size</a>.<a href="index.size.attributes.html">attributes</a>.removed</h1>
<p class="desc">Removing content attributes reverts to default size</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.size.missing">size.missing</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="60" id="c" width="120"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:</p><p><img alt="" class="output expected" id="expected" src="size.attributes.removed.png">
</p><ul id="d"></ul>
<script>
var t = async_test("Removing content attributes reverts to default size");
_addTest(function(canvas, ctx) {

_assertSame(canvas.width, 120, "canvas.width", "120");
canvas.removeAttribute('width');
_assertSame(canvas.width, 300, "canvas.width", "300");


});
</script>

<div id="log"></div>
</body></html>
